requirements.txt
jsonlines==2.0.0
Pillow==8.4.0
Pillow==10.0.0
Requests==2.31.0
